feat: refactor forgot password functionality; implement new form component and improve UI; enhance user session handling in login form
