+++ app/(auth)/forgot-password/new-password/page.tsx
