--- app/forgot-password/layout.tsx
@@ -1,3 +1,0 @@
-export default function ForgotPasswordLayout({ children }: { children: React.ReactNode }) {
-    return <div className="min-h-screen bg-yellow-50">{children}</div>;
-}
